# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Oracle Factory server.",
        "title": "Oracle Factory API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/webServiceListeners": {
            "get": {
                "description": "Retrieves all web service listeners for a user. This service will be called by the frontend to retrieve all web service listeners of the user signed in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServiceListener"
                ],
                "summary": "Retrieves all web service listeners for a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a web service listeners for a user. This service will be called by the frontend to when a user filled out the web service listener form.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServiceListener"
                ],
                "summary": "Creates a web service listeners for a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webServiceListeners/{webServiceListenerID}": {
            "get": {
                "description": "Retrieves the web service listener specified. This service will be called by the frontend to retrieve a specific listeners of the user signed in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServiceListener"
                ],
                "summary": "Retrieves a web service listener for a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the web service listener to send data to.",
                        "name": "webServiceListenerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webServiceListeners/{webServiceListenerID}/events": {
            "post": {
                "description": "Handles the event send from a web service provider. This endpoint will be called from an external service, that provides data to the artifact.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServiceListener"
                ],
                "summary": "Handles the event send from a web service provider",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the web service listener to send data to.",
                        "name": "webServiceListenerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webServicePublishers": {
            "get": {
                "description": "Retrieves all web service publishers for a user. This service will be called by the frontend to retrieve all web service publishers of the user signed in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServicePublisher"
                ],
                "summary": "Retrieves all web service publishers for a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a web service publishers for a user. This service will be called by the frontend to when a user filled out the web service publisher form.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServicePublisher"
                ],
                "summary": "Creates a web service publishers for a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webServicePublishers/{webServicePublisherID}": {
            "get": {
                "description": "Retrieves the web service publisher specified. This service will be called by the frontend to retrieve a specific publishers of the user signed in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webServicePublisher"
                ],
                "summary": "Retrieves a web service publisher for a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the web service publisher to send data to.",
                        "name": "webServicePublisherID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
...
    }
}